dsmca/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
PyYAML>=6.0.1
requests>=2.31.0
python-multipart>=0.0.6
pycryptodome>=3.18.0
python-jose>=3.3.0
typing-extensions>=4.8.0
websocket-client>=1.6.4 